+++ build.gradle
@@ -22,7 +22,7 @@ repositories {
 dependencies {
     compile 'com.google.guava:guava:24.0-jre'
     compile 'mysql:mysql-connector-java:5.1.25'
-    compile 'com.opuscapita.peppol:peppol-commons:0.0.32'
+    compile 'com.opuscapita.peppol:peppol-commons:0.0.33'
 
     compile 'org.springframework.boot:spring-boot-starter-thymeleaf:2.1.3.RELEASE'
     compile 'org.springframework.boot:spring-boot-starter-data-jpa:2.1.3.RELEASE'
