--- src/main/client/components/PeppolMonitorDetail/index.js
@@ -1,3 +1,0 @@
-import PeppolMonitorDetail from './PeppolMonitorDetail.react.js';
-
-export default PeppolMonitorDetail;
