+++ src/main/client/components/ProcessDetail/index.js
@@ -1,0 +1,3 @@
+import ProcessDetail from './ProcessDetail.react.js';
+
+export default ProcessDetail;
