+++ src/main/client/components/ProcessTable/ProcessTable.css
@@ -1,0 +1,30 @@
+.process-search .react-select {
+    min-width: 150px;
+    vertical-align: middle;
+}
+
+.process-list-table .label + .label {
+    margin-left: 3px;
+}
+
+.process-list-table .rt-td {
+    display: flex;
+    align-items: center;
+}
+
+.process-list-table .well {
+    margin: 0;
+    padding: 5px;
+    -webkit-border-radius: 5px;
+    -moz-border-radius: 5px;
+    border-radius: 5px;
+}
+
+.process-list-table .sticky {
+    pointer-events: auto !important;
+}
+
+.process-list-table .sticky:hover {
+    visibility: visible !important;
+    opacity: 1 !important;
+}
