+++ src/main/client/components/ProcessTable/index.js
@@ -1,0 +1,3 @@
+import ProcessTable from './ProcessTable.react.js';
+
+export default ProcessTable;
