+++ src/main/client/local.js
@@ -1,7 +1,6 @@
 'use strict';
 import 'babel-polyfill';
 import PeppolMonitor from './components/PeppolMonitor';
-import PeppolMonitorDetail from './components/PeppolMonitorDetail';
 
 const React = require('react');
 const ReactDOM = require('react-dom');
@@ -9,7 +8,6 @@ const ReactDOM = require('react-dom');
 const Local = () => (
     <div className={'wrapper'}>
         <PeppolMonitor/>
-        {/*<PeppolMonitorDetail messageId={2}/>*/}
     </div>
 );
